--- e2e/logic/api/responses/LoginResponse.ts
@@ -1,9 +1,0 @@
-export interface AuthCredentialsResponse {
-    credentials: {
-        id: string;
-        name: string;
-        type: string;
-        signinUrl: string;
-        callbackUrl: string;
-    };
-}
