--- e2e/logic/api/responses/addGraphResponse.ts
@@ -1,8 +1,0 @@
-export interface AddGraphResponse {
-  result: {
-    metadata: string[];
-    data: {
-      [key: string]: number;
-    }[];
-  };
-}
