--- e2e/logic/api/responses/addSchemaResponse.ts
@@ -1,6 +1,0 @@
-export interface AddSchemaResponse {
-    result: {
-        metadata: string[];
-        data: Array<Record<string, number>>;
-    };
-}
