+++ e2e/logic/api/responses/authResponses.ts
@@ -1,0 +1,13 @@
+export interface AuthCredentialsResponse {
+  credentials: {
+    id: string;
+    name: string;
+    type: string;
+    signinUrl: string;
+    callbackUrl: string;
+  };
+}
+
+export interface LogoutResponse {
+  url: string;
+}
