--- e2e/logic/api/responses/changeGraphNameResponse.ts
@@ -1,4 +1,0 @@
-export interface ChangeGraphNameResponse {
-    data: boolean;
-}
-
