--- e2e/logic/api/responses/createUsersResponse.ts
@@ -1,3 +1,0 @@
-export interface CreateUsersResponse {
-  message: string;
-}
