--- e2e/logic/api/responses/deleteUsersResponse.ts
@@ -1,3 +1,0 @@
-export interface DeleteUsersResponse {
-  message: string;
-}
