--- e2e/logic/api/responses/duplicateGraph.ts
@@ -1,3 +1,0 @@
-export interface DuplicateGraphresponse {
-    success: string;
-}
