--- e2e/logic/api/responses/getGraphsResponse.ts
@@ -1,3 +1,0 @@
-export interface GetGraphsResponse {
-    opts: string[];
-}
