--- e2e/logic/api/responses/getSchemaResponse.ts
@@ -1,3 +1,0 @@
-export interface SchemaListResponse {
-    opts: string[];
-}
