--- e2e/logic/api/responses/getSettingsRoleValue.ts
@@ -1,3 +1,0 @@
-export interface GetSettingsRoleValue {
-    config: [string, number];
-}
