--- e2e/logic/api/responses/getUsersResponse.ts
@@ -1,7 +1,0 @@
-export interface GetUsersResponse {
-  result: {
-    username: string;
-    role: string;
-    checked: boolean;
-  }[];
-}
