--- e2e/logic/api/responses/graphAttributeResponse.ts
@@ -1,5 +1,0 @@
-export interface GraphAttributeResponse {
-    result: {
-        metadata: string[];
-    };
-}
