--- e2e/logic/api/responses/graphCountResponse.ts
@@ -1,6 +1,0 @@
-export interface GraphCountResponse {
-  result: {
-      nodes: number;
-      edges: number;
-  };
-}
