--- e2e/logic/api/responses/graphNodeResponse.ts
@@ -1,3 +1,0 @@
-export interface GraphNodeResponse {
-    message: string;
-}
