--- e2e/logic/api/responses/logoutResponse.ts
@@ -1,3 +1,0 @@
-export interface LogoutResponse {
-    url: string;
-}
