--- e2e/logic/api/responses/modifySettingsRoleResponse.ts
@@ -1,3 +1,0 @@
-export interface ModifySettingsRoleResponse {
-  config: string;
-}
