--- e2e/logic/api/responses/removeGraphResponse.ts
@@ -1,3 +1,0 @@
-export interface RemoveGraphResponse {
-  message: string;
-}
