--- e2e/logic/api/responses/removeSchemaResponse.ts
@@ -1,3 +1,0 @@
-export interface RemoveSchemaResponse {
-    message: string;
-}
