--- e2e/logic/api/responses/runQueryResponse.ts
@@ -1,12 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-export interface RunQueryResponse {
-    metadata: {
-        "Labels added": number;
-        "Nodes created": number;
-        "Relationships created": number;
-        "Cached execution": number;
-        "Query internal execution time": string;
-    };
-    data: Array<Record<string, any>>;
-}
