+++ e2e/logic/api/responses/schemaResponses.ts
@@ -1,0 +1,14 @@
+export interface AddSchemaResponse {
+  result: {
+    metadata: string[];
+    data: Array<Record<string, number>>;
+  };
+}
+
+export interface SchemaListResponse {
+  opts: string[];
+}
+
+export interface RemoveSchemaResponse {
+  message: string;
+}
