+++ e2e/logic/api/responses/settingsResponses.ts
@@ -1,0 +1,7 @@
+export interface ModifySettingsRoleResponse {
+  config: string;
+}
+
+export interface GetSettingsRoleValue {
+  config: [string, number];
+}
