+++ e2e/logic/api/responses/userResponses.ts
@@ -1,0 +1,15 @@
+export interface GetUsersResponse {
+  result: {
+    username: string;
+    role: string;
+    checked: boolean;
+  }[];
+}
+
+export interface CreateUsersResponse {
+  message: string;
+}
+
+export interface DeleteUsersResponse {
+  message: string;
+}
